refactor: merge electricity and ceei dashboards codes to (app) main folder
