+++ .changeset/seven-frogs-own.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+refactor: merge electricity and ceei dashboards codes to (app) main folder
