+++ sites/geohub/src/routes/(app)/dashboards/ceei/components/DropdownSearch.svelte
