+++ sites/geohub/src/routes/(app)/dashboards/ceei/components/PopupTable.svelte
