+++ sites/geohub/src/routes/(app)/dashboards/ceei/stores/index.ts
