+++ sites/geohub/src/routes/(app)/dashboards/ceei/utils/layerHelper.ts
