+++ sites/geohub/src/routes/(app)/dashboards/electricity/components/AnalyzeBivariate.svelte
