+++ sites/geohub/src/routes/(app)/dashboards/electricity/components/Charts.svelte
