+++ sites/geohub/src/routes/(app)/dashboards/electricity/components/ElectricityControl.svelte
