+++ sites/geohub/src/routes/(app)/dashboards/electricity/components/ExploreEvolution.svelte
