+++ sites/geohub/src/routes/(app)/dashboards/electricity/components/IntroductionPanel.svelte
