+++ sites/geohub/src/routes/(app)/dashboards/electricity/components/TimeSlider.svelte
