+++ sites/geohub/src/routes/(app)/dashboards/electricity/components/TimeSliderControl.svelte
