+++ sites/geohub/src/routes/(app)/dashboards/electricity/constansts.ts
@@ -1,6 +1,6 @@
 // must specify dataset ID of GeoHub api
 
-import type { DashBoardDataset } from './+page.svelte';
+import type { DashBoardDataset } from './+page@.svelte';
 
 // token and URL of datasets will be retrieved from /api/datasets/{id} API
 export const ELECTRICITY_DATASETS: { [key: string]: DashBoardDataset[] } = {
