+++ sites/geohub/src/routes/(app)/dashboards/electricity/stores/electricityDataType.ts
