+++ sites/geohub/src/routes/(app)/dashboards/electricity/stores/index.ts
