+++ sites/geohub/src/routes/(app)/dashboards/electricity/utils/adminLayer.ts
