--- sites/geohub/src/routes/(map)/dashboards/+layout.svelte
@@ -1,32 +1,0 @@
-<script lang="ts">
-	import Header from '$components/header/Header.svelte';
-	import { createHeaderHeightStore, HEADER_HEIGHT_CONTEXT_KEY } from '$stores';
-	import { addProtocol } from 'maplibre-gl';
-	import * as pmtiles from 'pmtiles';
-	import { onMount, setContext } from 'svelte';
-	interface Props {
-		children?: import('svelte').Snippet;
-	}
-
-	let { children }: Props = $props();
-
-	const headerHeightStore = createHeaderHeightStore();
-	setContext(HEADER_HEIGHT_CONTEXT_KEY, headerHeightStore);
-
-	onMount(() => {
-		let protocol = new pmtiles.Protocol();
-		addProtocol('pmtiles', protocol.tile);
-	});
-</script>
-
-<svelte:head>
-	<style type="text/css">
-		html {
-			overflow-y: hidden !important;
-		}
-	</style>
-</svelte:head>
-
-<Header isPositionFixed={true} />
-
-{@render children?.()}
